Update myLittleVideoGame.py
I added some changes, you can switch out backgrounds.  Also, regarding the music, I use try except so that program will not crash if using Linux.  So at this time, background music plays well in Linux, as well as playSoundAndPauseWhilePlaying method.  However, playSoundandKeepGoing does not work in Linux, but the game will keep going.  I added a "RotatingCannon" class, because it was required for one of my games.  In LittleObjects, I refreshed the window right after initialization of the instance of a LittleObject.  I intend to share an example of the use.  Or you can go to mylittlevideogame.com, and there is a project with a Windows installer under "Base Invader Trilogy"